refactor: mypagelayout 컴포넌트 나눔, supabase 테이블 명칭 변경 userinfo -> user_info, / chore: prettier setting 코드컨벤션 셋팅으로 설정
